requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
colorama==0.4.4
defusedxml==0.7.1
Django==4.0
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-crispy-forms==1.13.0
django-embed-video==1.4.0
django-js-asset==1.2.2
djangorestframework==3.13.1
	djangorestframework-jwt==1.11.0
	djangorestframework-simplejwt==5.0.0
httpie==2.6.0
idna==3.3
mysqlclient @ file:///C:/Users/Kiranp/Desktop/python/post/blog_project/mysqlclient-1.4.6-cp39-cp39-win_amd64.whl
Pygments==2.11.2
	PyJWT==2.3.0
 PySocks==1.7.1
 pytz==2021.3
requests==2.26.0
requests-toolbelt==0.9.1
sqlparse==0.4.2
 tzdata==2021.5
 urllib3==1.26.7
